+++ packages/react/CHANGELOG.md
@@ -2,11 +2,20 @@
 
 Nimbus is an open-source Design System created by Tiendanube / Nuvesmhop's team to empower and enhance more stories every day, with simplicity, accessibility, consistency and performance.
 
-## 2025-12-23 `1.29.0`
+## 2026-02-05 `1.30.0`
 
 #### 🎉 New features
 
 - Added `paddingBottom` prop to `DataList.Row` sub-component to allow custom padding bottom definition to rows. ([#141](https://github.com/TiendaNube/nimbus-patterns/pull/141) by [@joacotornello](https://github.com/joacotornello))
+
+## 2026-02-03 `1.29.0`
+
+#### 🎉 New features
+
+- Added `DataTable.Dropdown` component. ([#141](https://github.com/TiendaNube/nimbus-patterns/pull/141) by [@joacotornello](https://github.com/joacotornello))
+- Added `DataTable.DropdownAction` component. ([#141](https://github.com/TiendaNube/nimbus-patterns/pull/141) by [@joacotornello](https://github.com/joacotornello))
+- Added `DataTable.DropdownDivider` component. ([#141](https://github.com/TiendaNube/nimbus-patterns/pull/141) by [@joacotornello](https://github.com/joacotornello))
+- Added `DataTable.DropdownSection` component. ([#141](https://github.com/TiendaNube/nimbus-patterns/pull/141) by [@joacotornello](https://github.com/joacotornello))
 
 ## 2026-01-22 `1.28.1`